Updated Firebase Realtime Database rules
We decide to utilize Firebase Realtime Database rules to enforce some of the
designed principles. Since this update, the firebase will refuse resetting any
decayed status again to non-decayed status, unless the server is reinitialized
with "--init" option.

This is a preparation for future modification to store encrypted credential at
Firebase.