Support encrypted config file.
YAML config file can be encrypted with gpg. Upon starting server, supply
a `.gpg` config file, the program will call gpg for decryption. This is a
small improvement on security, so that important credentials(for Firebase
and Telegram) will not be saved at disk in plain, as these can be used for
extracting Remote Encryption Key(REK) from Firebase.

It's planned to implement a "keeper" role in VNMK system, so that multiple
REKs(distributed at very different servers--different service providers
and under different jurisdictions) can be used simultaneously.